docs/requirements.txt
GitPython
numpydoc
sphinx
sphinx-copybutton
sphinx-codeautolink
